feat: add content/skeleton loader to product card
