+++ package.json
@@ -19,8 +19,8 @@
     "eslint-config-next": "13.1.6",
     "next": "13.1.6",
     "react": "18.2.0",
+    "react-content-loader": "^6.2.0",
     "react-dom": "18.2.0",
-    "react-loading-skeleton": "^3.1.0",
     "react-redux": "^8.0.5",
     "styled-components": "^5.3.6",
     "typescript": "4.9.4"
